Remove unused primitives from hook
